added index to each module and fixed imports accordingly
The goal of this change is to streamline imports so that each module can communicate with another only through its index. This sort of interface makes each module more indipendent and enables it to have more freedom of movement. However, since ES6 imports cannot be scoped and can be accessed from anywhere, this requires the developer to actively avoid importing JS modules contained in @/modules other than the index. Yet, in order to enforce this, lint rules could be added.